+++ packages/server/index.ts
@@ -1,4 +1,5 @@
 export * from './src/startServer';
+export * from './src/GracefulShutdown';
 export * from './src/envCache';
 export * from './src/Request';
 export * from './src/SocketIOServerRepo';